display -> tick 変更, Fps extends Polling
